feat: move svelte interface to 'src/interface'
