+++ src/interface/components/ColourPicker.svelte
