+++ src/interface/components/PickerSwatch.svelte
