+++ src/interface/components/SkeletonLoader.svelte
