+++ src/interface/components/Slider.svelte
