+++ src/interface/components/Spinner.svelte
