+++ src/interface/components/TabbedContainer.svelte
