+++ src/interface/components/store/FilterPanel.svelte
