+++ src/interface/components/themeCreator/divider.svelte
