+++ src/interface/components/themes/BackgroundItem.svelte
