+++ src/interface/components/themes/BackgroundUploader.svelte
