+++ src/interface/components/utils/ReactAdapter.svelte
