+++ src/interface/hooks/backgroundState.svelte.ts
