+++ src/interface/main.d.ts
