+++ src/interface/utils/standalone.svelte.ts
