+++ src/interface/utils/themeImageHandlers.ts
