+++ src/manifests/manifest.json
@@ -10,7 +10,7 @@
   },
   "action": {
     "browser_style": true,
-    "default_popup": "svelte-interface/index.html#settings",
+    "default_popup": "interface/index.html#settings",
     "default_icon": {
       "32": "resources/icons/icon-32.png",
       "48": "resources/icons/icon-48.png",
