+++ vite.config.ts
@@ -60,7 +60,7 @@ export default defineConfig({
     minify: false,
     rollupOptions: {
       input: {
-        settings: join(__dirname, 'src', 'svelte-interface', 'index.html')
+        settings: join(__dirname, 'src', 'interface', 'index.html')
       }
     }
   }
